Fix warnings of conversion

+++ test/case/channel_mem_test.cpp
@@ -81,7 +81,7 @@ CASE_TEST(channel, mem_attach_with_invalid_host_size) {
   CASE_EXPECT_EQ(sizeof(size_t), mem_info_get_host_size(channel));
   CASE_EXPECT_EQ(0, mem_info_get_host_size(nullptr));
 
-  (*reinterpret_cast<uint16_t *>(buffer + 20)) = sizeof(size_t) / 2;
+  (*reinterpret_cast<uint16_t *>(buffer + 20)) = static_cast<uint16_t>(sizeof(size_t) / 2);
   CASE_EXPECT_EQ(EN_ATBUS_ERR_CHANNEL_ARCH_SIZE_T_MISMATCH, mem_attach(buffer, buffer_len, &channel, nullptr));
 
   delete[] buffer;
